i genuinely have NO idea why, but i have to undo my flash-of-unpositioned-model fix in order for a nested model to work in the project component????
there must be something going on with a subcomponent mounting that causes extra renders and runs of that firstRun stuff?
